+++ changelog.md
@@ -18,6 +18,7 @@
 * #216 Better return request support
 * #90 Directly set order status to completed
 * #413 Orders: Add a PDF export\download of selected orders
+* #69 Award reward points for product reviews
 
 ###Improvements###
 * Task Scheduler: